Actualizacion de rutas
