+++ BLOGS/templates/crearEntrada.html
@@ -8,7 +8,7 @@
     <!-- Iconos -->
     <link href="https://unpkg.com/ionicons@4.5.10-0/dist/css/ionicons.min.css" rel="stylesheet">
     <!-- Mi CSS -->
-    <link href="styles.css" rel="stylesheet" type="text/css">
+    <link rel="stylesheet" type="text/css" href="{{url_for('static',filename='styles.css') }}">
     <title>Nuevo Blog</title>
 </head>
 <body>
