+++ BLOGS/templates/login.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+    <!-- Mi CSS -->
+    <link rel="stylesheet" type="text/css" href="{{url_for('static',filename='styles.css') }}">
+    
+</head>
+<body>
+    <div>
+        <form class="login" action="/crearEntrada" method="POST"> 
+            <div class="form-group row">
+                <label for="userEmail" class="col-sm-2 col-form-label">Usuario:</label>
+                <div class="col-sm-10">
+                    <input type="email" class="form-control" id="userEmail" name="usuario" aria-describedby="emailHelp" placeholder="Ingresa tu correo">
+                </div>
+            </div>
+            <br>
+            <div class="form-group row">
+                <label for="userPassword" class="col-sm-2 col-form-label">Contraseña:</label>
+                <div class="col-sm-10">
+                    <input type="password" class="form-control" id="userPassword" name="password" placeholder="Ingresa tu contraseña">
+                </div>
+            </div>
+            <center>
+            <br>
+            <button type="submit" class="btn btn-primary boton-ingresar">Ingresar</button>
+            <br><br>
+            <a href=>Crear usuario</a>
+            </center>
+        
+            
+        </form>
+        </div>
+        <!-- Option 1: jQuery and Bootstrap Bundle (includes Popper) -->
+        <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ho+j7jyWK8fNQe+A12Hb8AhRq26LrZ/JpcUGGOn+Y7RsweNrtN/tE3MoK7ZeZDyx" crossorigin="anonymous"></script>
+</body>
+</html>
